Check for Modules Existence Before Deleting
We have some weird issue at PayPal where we see this `module` variable as `undefined` and it will blow up, but only when we have the `webpack` `eslint` plugin enabled. I don't know why it is, but this fixes the issue, assume `eslint` messes with the `require.cache` or something? A bit weird, but wondering if you'd consider this change (or not).